challenge3/requirements.txt
openai
python-dotenv
azure-identity
azure-search-documents
azure-core